syncer: move mutators into spec/mutators
